console: Move device bulk creation up in routing hierarchy
