+++ pkg/webui/console/containers/devices-table/devices-table.styl
@@ -27,5 +27,5 @@
     align-items: center
     max-width: 30rem
 
-.add-button
-  margin-left: $cs.xs
+.import-devices
+  margin-right: 0
