+++ pkg/webui/console/containers/fetch-table/fetch-table.styl
@@ -31,9 +31,6 @@
     max-width: 26rem
     min-height: 3.8rem
 
-.add-button
-  margin-left: $cs.xs
-
 .page
   margin-top: $ls.s
