+++ pkg/webui/console/views/device-add-single/device-add-single.styl
@@ -16,11 +16,6 @@
   text-margin-top($ls.m)
   text-margin-bottom($ls.m)
 
-.bulk-creation
-  justify-content: flex-end
-  display: flex
-  align-items: center
-
 .form
   +media-query($bp.s)
     order: 2
